refactor: dps score panel
