+++ src/lib/characterPreview/CharacterPreview.tsx
@@ -1,30 +1,30 @@
-import { EditOutlined, SettingOutlined, SwapOutlined, SyncOutlined } from '@ant-design/icons'
-import { Button, Card, Flex, Image, Segmented, theme, Typography } from 'antd'
+import { EditOutlined } from '@ant-design/icons'
+import { Button, Flex, Image, Segmented, theme, Typography } from 'antd'
 import CharacterCustomPortrait from 'lib/characterPreview/CharacterCustomPortrait'
-import { ShowcaseRelicPreview, ShowcaseSource } from 'lib/characterPreview/CharacterPreviewComponents'
+import { showcaseButtonStyle, showcaseDropShadowFilter, showcaseOutline, showcaseShadow, ShowcaseSource } from 'lib/characterPreview/CharacterPreviewComponents'
 import { getArtistName, getPreviewRelics, presetTeamSelectionDisplay, showcaseIsInactive } from 'lib/characterPreview/characterPreviewController'
 import { CharacterCardCombatStats, CharacterCardScoringStatUpgrades, CharacterScoringSummary } from 'lib/characterPreview/CharacterScoringSummary'
 import { CharacterStatSummary } from 'lib/characterPreview/CharacterStatSummary'
 
 import Rarity from 'lib/characterPreview/Rarity'
+import { ShowcaseDpsScoreHeader, ShowcaseDpsScorePanel } from 'lib/characterPreview/ShowcaseDpsScore'
+import { ShowcaseRelicsPanel } from 'lib/characterPreview/ShowcaseRelicsPanel'
 import StatText from 'lib/characterPreview/StatText'
-import { CHARACTER_SCORE, COMBAT_STATS, CUSTOM_TEAM, DAMAGE_UPGRADES, DEFAULT_TEAM, ElementToDamage, SETTINGS_TEAM, SIMULATION_SCORE } from 'lib/constants/constants'
+import { CHARACTER_SCORE, COMBAT_STATS, CUSTOM_TEAM, DAMAGE_UPGRADES, DEFAULT_TEAM, ElementToDamage, SIMULATION_SCORE } from 'lib/constants/constants'
 import { SavedSessionKeys } from 'lib/constants/constantsSession'
 import { defaultGap, innerW, lcInnerH, lcInnerW, lcParentH, lcParentW, middleColumnWidth, parentH, parentW } from 'lib/constants/constantsUi'
 import { Message } from 'lib/interactions/message'
 import { calculateBuild } from 'lib/optimization/calculateBuild'
-import CharacterModal from 'lib/overlays/modals/CharacterModal'
 import EditImageModal from 'lib/overlays/modals/EditImageModal'
 import RelicModal from 'lib/overlays/modals/RelicModal'
 import { RelicModalController } from 'lib/overlays/modals/relicModalController'
 import { RelicFilters } from 'lib/relics/relicFilters'
 import { StatCalculator } from 'lib/relics/statCalculator'
 import { Assets } from 'lib/rendering/assets'
-import { getSimScoreGrade, scoreCharacterSimulation } from 'lib/scoring/characterScorer'
+import { scoreCharacterSimulation } from 'lib/scoring/characterScorer'
 import { DB } from 'lib/state/db'
 import { SaveState } from 'lib/state/saveState'
 import { OptimizerTabController } from 'lib/tabs/tabOptimizer/optimizerTabController'
-import { HeaderText } from 'lib/ui/HeaderText'
 import { LoadingBlurredImage } from 'lib/ui/LoadingBlurredImage'
 import { Utils } from 'lib/utils/utils'
 import React, { useEffect, useRef, useState } from 'react'
@@ -36,18 +36,9 @@ import { Relic } from 'types/relic'
 const { useToken } = theme
 const { Text } = Typography
 
-const outline = 'rgb(255 255 255 / 40%) solid 1px'
-const shadow = 'rgba(0, 0, 0, 0.5) 1px 1px 1px 1px'
-const filter = 'drop-shadow(rgb(0, 0, 0) 1px 1px 3px)'
-const buttonStyle = {
-  flex: 'auto',
-  opacity: 0,
-  transition: 'opacity 0.3s ease',
-  visibility: 'hidden',
-}
-
 // This is hardcoded for the screenshot-to-clipboard util. Probably want a better way to do this if we ever change background colors
 export function CharacterPreview(props: {
+  id: string
   source: ShowcaseSource
   character: Character
   setOriginalCharacterModalOpen: (open: boolean) => void
@@ -68,18 +59,18 @@ export function CharacterPreview(props: {
   const { token } = useToken()
 
   const relicsById = window.store((s) => s.relicsById)
-  const [selectedRelic, setSelectedRelic] = useState<Relic | null>(null)
+  const [selectedRelic, setSelectedRelic] = useState<Relic | undefined>()
   const [editModalOpen, setEditModalOpen] = useState(false)
   const [addModalOpen, setAddModalOpen] = useState(false)
   const [editPortraitModalOpen, setEditPortraitModalOpen] = useState(false)
-  const [customPortrait, setCustomPortrait] = useState<CustomImageConfig | null>(null) // <null | CustomImageConfig>
+  const [customPortrait, setCustomPortrait] = useState<CustomImageConfig | undefined>() // <null | CustomImageConfig>
   const [teamSelection, setTeamSelection] = useState(DEFAULT_TEAM)
   const [scoringType, setScoringType] = useState(SIMULATION_SCORE)
   const [combatScoreDetails, setCombatScoreDetails] = useState(COMBAT_STATS)
   const [isCharacterModalOpen, setCharacterModalOpen] = useState(false)
-  const [characterModalInitialCharacter, setCharacterModalInitialCharacter] = useState()
-  const [selectedTeammateIndex, setSelectedTeammateIndex] = useState()
-  const [redrawTeammates, setRedrawTeammates] = useState()
+  const [characterModalInitialCharacter, setCharacterModalInitialCharacter] = useState<Character | undefined>()
+  const [selectedTeammateIndex, setSelectedTeammateIndex] = useState<number | undefined>()
+  const [redrawTeammates, setRedrawTeammates] = useState<number>(0)
   const activeKey = window.store((s) => s.activeKey)
 
   // We need to track the previously selected character in order to know which state to put the sim team in.
@@ -239,205 +230,6 @@ export function CharacterPreview(props: {
   // Since the lc takes some space, we want to zoom the portrait out
   const tempInnerW = simScoringResult ? 950 : innerW
 
-  // Teammate character modal OK
-  function onCharacterModalOk(form) {
-    if (!form.characterId) {
-      return Message.error(t('CharacterPreview.Messages.NoSelectedCharacter')/* No selected character */)
-    }
-    if (!form.lightCone) {
-      return Message.error(t('CharacterPreview.Messages.NoSelectedCharacter')/* No selected character */)
-    }
-
-    const scoringMetadata = Utils.clone(DB.getScoringMetadata(characterId))
-    const simulation = scoringMetadata.simulation
-
-    simulation.teammates[selectedTeammateIndex] = form
-
-    DB.updateSimulationScoreOverrides(characterId, simulation)
-    setRedrawTeammates(Utils.randomId())
-
-    setTeamSelection(CUSTOM_TEAM)
-  }
-
-  function ScoreHeader(props) {
-    const result = props.result
-    const verified = Object.values(props.relics).filter((x) => x?.verified).length == 6
-
-    const textStyle = {
-      fontSize: 17,
-      fontWeight: '600',
-      textAlign: 'center',
-      color: 'rgb(225, 165, 100)',
-      height: 23,
-      whiteSpace: 'nowrap',
-    }
-
-    const textDisplay = (
-      <Flex align='center' vertical style={{ marginBottom: 4, paddingTop: 3, paddingBottom: 3 }}>
-        <StatText style={textStyle}>
-          {t('CharacterPreview.ScoreHeader.Title')/* Combat Sim */}
-        </StatText>
-        <StatText style={textStyle}>
-          {
-            t(
-              'CharacterPreview.ScoreHeader.Score',
-              {
-                score: Utils.truncate10ths(Math.max(0, result.percent * 100)).toFixed(1),
-                grade: getSimScoreGrade(result.percent, verified),
-              },
-            )
-            /* DPS Score {{score}}% {{grade}} */
-          }
-        </StatText>
-      </Flex>
-    )
-
-    return (
-      <Flex vertical>
-        {textDisplay}
-      </Flex>
-    )
-  }
-
-  function ScoreFooter(props) {
-    const tabsDisplay = (
-      <Segmented
-        style={{ marginLeft: 10, marginRight: 10, marginTop: 2, marginBottom: 0, alignItems: 'center' }}
-        onChange={(selection) => {
-          if (selection == SETTINGS_TEAM) {
-            window.modalApi.info({
-              icon: null,
-              width: 400,
-              okText: t('common:Ok'),
-              maskClosable: true,
-              content: (
-                <div style={{ width: '100%' }}>
-                  <Flex vertical gap={10}>
-                    <HeaderText>{t('modals:ScoreFooter.ModalTitle')/* Combat sim scoring settings */}</HeaderText>
-                    <Button
-                      icon={<SyncOutlined/>}
-                      onClick={() => {
-                        const characterMetadata = Utils.clone(DB.getMetadata().characters[character.id])
-                        const simulation = characterMetadata.scoringMetadata.simulation
-
-                        DB.updateSimulationScoreOverrides(character.id, simulation)
-
-                        setTeamSelection(DEFAULT_TEAM)
-                        setRedrawTeammates(Math.random())
-
-                        Message.success(t('modals:ScoreFooter.ResetSuccessMsg')/* Reset to default teams */)
-                      }}
-                    >
-                      {t('modals:ScoreFooter.ResetButtonText')/* Reset custom team to default */}
-                    </Button>
-                    <Button
-                      icon={<SwapOutlined/>}
-                      onClick={() => {
-                        const characterMetadata = Utils.clone(DB.getScoringMetadata(character.id))
-                        const simulation = characterMetadata.simulation
-
-                        for (const teammate of simulation.teammates) {
-                          const form = DB.getCharacterById(teammate.characterId)?.form
-                          if (form == null) continue
-
-                          teammate.characterEidolon = form.characterEidolon
-                          if (form.lightCone) {
-                            teammate.lightCone = form.lightCone
-                            teammate.lightConeSuperimposition = form.lightConeSuperimposition || 1
-                          }
-                        }
-
-                        DB.updateSimulationScoreOverrides(character.id, simulation)
-                        setTeamSelection(CUSTOM_TEAM)
-                        setRedrawTeammates(Math.random())
-
-                        Message.success(t('modals:ScoreFooter.SyncSuccessMsg')/* Synced teammates */)
-                      }}
-                    >
-                      {t('modals:ScoreFooter.SyncButtonText')/* Sync imported eidolons / light cones */}
-                    </Button>
-                  </Flex>
-                </div>
-              ),
-            })
-          } else {
-            setTeamSelection(selection)
-          }
-        }}
-        value={teamSelection}
-        block
-        options={[
-          {
-            value: DEFAULT_TEAM,
-            label: t('modals:ScoreFooter.TeamOptions.Default')/* Default */,
-          },
-          {
-            label: (
-              <SettingOutlined/>
-            ),
-            value: SETTINGS_TEAM,
-            className: 'short-segmented',
-          },
-          {
-            value: CUSTOM_TEAM,
-            label: t('modals:ScoreFooter.TeamOptions.Custom')/* Custom */,
-          },
-        ]}
-      />
-    )
-
-    return (
-      <Flex vertical style={{}} gap={2}>
-        <CharacterModal
-          onOk={onCharacterModalOk}
-          open={isCharacterModalOpen}
-          setOpen={setCharacterModalOpen}
-          initialCharacter={characterModalInitialCharacter}
-        />
-        {tabsDisplay}
-      </Flex>
-    )
-  }
-
-  function CharacterPreviewScoringTeammate(props) {
-    const { result, index, setCharacterModalOpen, setSelectedTeammateIndex } = props
-    const teammate = result.simulationMetadata.teammates[index]
-    const iconSize = 64
-    return (
-      <Card.Grid
-        style={gridStyle} hoverable={true}
-        onClick={() => {
-          setCharacterModalInitialCharacter({ form: teammate })
-          setCharacterModalOpen(true)
-
-          setSelectedTeammateIndex(index)
-        }}
-        className='custom-grid'
-      >
-        <Flex vertical align='center' gap={0}>
-          <img
-            src={Assets.getCharacterAvatarById(teammate.characterId)}
-            style={{
-              height: iconSize,
-              width: iconSize,
-              borderRadius: 40,
-              background: 'rgba(255, 255, 255, 0.1)',
-              boxShadow: '0 4px 30px rgba(0, 0, 0, 0.2)',
-              backdropFilter: 'blur(5px)',
-              outline: '1px solid rgba(255, 255, 255, 0.3)',
-            }}
-          />
-          <OverlayText text={t('common:EidolonNShort', { eidolon: teammate.characterEidolon })} top={-12}/>
-          <img src={Assets.getLightConeIconById(teammate.lightCone)} style={{ height: iconSize, marginTop: -3 }}/>
-          <OverlayText
-            text={t('common:SuperimpositionNShort', { superimposition: teammate.lightConeSuperimposition })}
-            top={-18}
-          />
-        </Flex>
-      </Card.Grid>
-    )
-  }
-
   return (
     <Flex vertical>
       <Flex vertical id={props.id} style={{ backgroundColor: backgroundColor }}>
@@ -449,20 +241,21 @@ export function CharacterPreview(props: {
           }}
         >
           <RelicModal
-            selectedRelic={selectedRelic} type='edit' onOk={onEditOk}
+            selectedRelic={selectedRelic}
+            type='edit'
+            onOk={onEditOk}
             setOpen={setEditModalOpen}
             open={editModalOpen}
           />
           <RelicModal
-            selectedRelic={selectedRelic} type='edit' onOk={onAddOk}
+            selectedRelic={selectedRelic}
+            type='edit'
+            onOk={onAddOk}
             setOpen={setAddModalOpen}
             open={addModalOpen}
           />
 
-          <Flex
-            vertical gap={12}
-            className='character-build-portrait'
-          >
+          <Flex vertical gap={12} className='character-build-portrait'>
             {source != ShowcaseSource.BUILDS_MODAL && (
               <div
                 style={{
@@ -471,8 +264,8 @@ export function CharacterPreview(props: {
                   overflow: 'hidden',
                   borderRadius: '8px',
                   marginRight: defaultGap,
-                  outline: outline,
-                  filter: filter,
+                  outline: showcaseOutline,
+                  filter: showcaseDropShadowFilter,
                   position: 'relative',
                 }}
               >
@@ -499,9 +292,7 @@ export function CharacterPreview(props: {
                 <Flex vertical style={{ width: 'max-content', marginLeft: 6, marginTop: 6 }} gap={7}>
                   {source != ShowcaseSource.SHOWCASE_TAB && (
                     <Button
-                      style={{
-                        ...buttonStyle,
-                      }}
+                      style={showcaseButtonStyle}
                       className='character-build-portrait-button'
                       icon={<EditOutlined/>}
                       onClick={() => {
@@ -516,9 +307,7 @@ export function CharacterPreview(props: {
                   )}
                   {source == ShowcaseSource.SHOWCASE_TAB && (
                     <Button
-                      style={{
-                        ...buttonStyle,
-                      }}
+                      style={showcaseButtonStyle}
                       className='character-build-portrait-button'
                       icon={<EditOutlined/>}
                       onClick={() => {
@@ -531,9 +320,7 @@ export function CharacterPreview(props: {
                     </Button>
                   )}
                   <Button
-                    style={{
-                      ...buttonStyle,
-                    }}
+                    style={showcaseButtonStyle}
                     className='character-build-portrait-button'
                     icon={<EditOutlined/>}
                     onClick={() => setEditPortraitModalOpen(true)}
@@ -613,8 +400,8 @@ export function CharacterPreview(props: {
                           whiteSpace: 'nowrap',
                           zIndex: 3,
                           textShadow: '0px 0px 10px black',
-                          outline: outline,
-                          boxShadow: shadow,
+                          outline: showcaseOutline,
+                          boxShadow: showcaseShadow,
                         }}
                       >
                         {`${t('common:SuperimpositionNShort', { superimposition: lightConeSuperimposition })} - ${lightConeName}`}
@@ -629,8 +416,8 @@ export function CharacterPreview(props: {
                       overflow: 'hidden',
                       zIndex: 2,
                       borderRadius: '8px',
-                      outline: outline,
-                      filter: filter,
+                      outline: showcaseOutline,
+                      filter: showcaseDropShadowFilter,
                       position: 'relative',
                     }}
                     onClick={() => {
@@ -697,52 +484,18 @@ export function CharacterPreview(props: {
                 />
                 {
                   simScoringResult
-                  && <ScoreHeader result={simScoringResult} relics={displayRelics}/>
+                  && <ShowcaseDpsScoreHeader result={simScoringResult} relics={displayRelics}/>
                 }
                 {
                   simScoringResult
                   && (
-                    <Flex
-                      vertical
-                    >
-                      <Card
-                        style={{
-                          backgroundColor: token.colorBgLayout,
-                          padding: '0px !important',
-                          body: {
-                            height: 500,
-                          },
-                        }}
-                        styles={{
-                          body: {
-                            padding: 0,
-                            borderRadius: 10,
-                          },
-                        }}
-                        className='teamSelectionCard'
-                        size='small'
-                        bordered={false}
-                      >
-                        <Flex justify='space-around' style={{ paddingTop: 0 }}>
-                          <CharacterPreviewScoringTeammate
-                            result={simScoringResult} index={0}
-                            setCharacterModalOpen={setCharacterModalOpen}
-                            setSelectedTeammateIndex={setSelectedTeammateIndex}
-                          />
-                          <CharacterPreviewScoringTeammate
-                            result={simScoringResult} index={1}
-                            setCharacterModalOpen={setCharacterModalOpen}
-                            setSelectedTeammateIndex={setSelectedTeammateIndex}
-                          />
-                          <CharacterPreviewScoringTeammate
-                            result={simScoringResult} index={2}
-                            setCharacterModalOpen={setCharacterModalOpen}
-                            setSelectedTeammateIndex={setSelectedTeammateIndex}
-                          />
-                        </Flex>
-                      </Card>
-                      <ScoreFooter result={simScoringResult}/>
-                    </Flex>
+                    <ShowcaseDpsScorePanel
+                      token={token}
+                      simScoringResult={simScoringResult}
+                      setCharacterModalOpen={setCharacterModalOpen}
+                      setSelectedTeammateIndex={setSelectedTeammateIndex}
+                      setCharacterModalInitialCharacter={setCharacterModalInitialCharacter}
+                    />
                   )
                 }
                 {
@@ -800,8 +553,8 @@ export function CharacterPreview(props: {
                           height: `${tempLcParentH}px`,
                           overflow: 'hidden',
                           borderRadius: '8px',
-                          outline: outline,
-                          filter: filter,
+                          outline: showcaseOutline,
+                          filter: showcaseDropShadowFilter,
                         }}
                         onClick={() => {
                           if (source == ShowcaseSource.SHOWCASE_TAB) {
@@ -828,7 +581,7 @@ export function CharacterPreview(props: {
               </Flex>
             </Flex>
 
-            <ShowcaseRelicPreview
+            <ShowcaseRelicsPanel
               setSelectedRelic={setSelectedRelic}
               setEditModalOpen={setEditModalOpen}
               setAddModalOpen={setAddModalOpen}
@@ -929,45 +682,4 @@ export function CharacterPreview(props: {
       )}
     </Flex>
   )
-}
-
-function OverlayText(props) {
-  const top = props.top
-  return (
-    <Flex
-      vertical
-      style={{
-        position: 'relative',
-        height: 0,
-        top: top,
-      }}
-      align='center'
-    >
-      <Text
-        style={{
-          position: 'absolute',
-          backgroundColor: 'rgb(0 0 0 / 75%)',
-          padding: '2px 14px',
-          borderRadius: 4,
-          fontSize: 12,
-          textOverflow: 'ellipsis',
-          whiteSpace: 'nowrap',
-          textShadow: '0px 0px 10px black',
-          outline: outline,
-          filter: filter,
-          lineHeight: '12px',
-        }}
-      >
-        {props.text}
-      </Text>
-    </Flex>
-  )
-}
-
-const gridStyle = {
-  width: '33.3333%',
-  textAlign: 'center',
-  padding: 1,
-  // border: 'none',
-  boxShadow: 'none',
 }
