+++ src/lib/characterPreview/ShowcaseDpsScore.tsx
@@ -1,0 +1,323 @@
+import { SettingOutlined, SwapOutlined, SyncOutlined } from '@ant-design/icons'
+import { Button, Card, Flex, Segmented } from 'antd'
+import type { GlobalToken } from 'antd/es/theme/interface'
+import { OverlayText } from 'lib/characterPreview/CharacterPreviewComponents'
+import StatText from 'lib/characterPreview/StatText'
+import { CUSTOM_TEAM, DEFAULT_TEAM, SETTINGS_TEAM } from 'lib/constants/constants'
+import { SingleRelicByPart } from 'lib/gpu/webgpuTypes'
+import { Message } from 'lib/interactions/message'
+import CharacterModal from 'lib/overlays/modals/CharacterModal'
+import { Assets } from 'lib/rendering/assets'
+import { getSimScoreGrade, SimulationScore } from 'lib/scoring/characterScorer'
+import DB from 'lib/state/db'
+import { HeaderText } from 'lib/ui/HeaderText'
+import { Utils } from 'lib/utils/utils'
+import React, { CSSProperties, useState } from 'react'
+import { useTranslation } from 'react-i18next'
+import { Character } from 'types/character'
+import { Form } from 'types/form'
+
+export function ShowcaseDpsScorePanel(props: {
+  token: GlobalToken
+  simScoringResult: SimulationScore
+  setCharacterModalOpen: (b: boolean) => void
+  setSelectedTeammateIndex: (i: number | undefined) => void
+  setCharacterModalInitialCharacter: (c: Character) => void
+}) {
+  const {
+    token,
+    simScoringResult,
+    setCharacterModalOpen,
+    setSelectedTeammateIndex,
+    setCharacterModalInitialCharacter,
+  } = props
+
+  return (
+    <Flex
+      vertical
+    >
+      <Card
+        style={{
+          backgroundColor: token.colorBgLayout,
+          padding: '0px !important',
+          body: {
+            height: 500,
+          },
+        }}
+        styles={{
+          body: {
+            padding: 0,
+            borderRadius: 10,
+          },
+        }}
+        className='teamSelectionCard'
+        size='small'
+        bordered={false}
+      >
+        <Flex justify='space-around' style={{ paddingTop: 0 }}>
+          <CharacterPreviewScoringTeammate
+            index={0}
+            result={simScoringResult}
+            setCharacterModalOpen={setCharacterModalOpen}
+            setSelectedTeammateIndex={setSelectedTeammateIndex}
+            setCharacterModalInitialCharacter={setCharacterModalInitialCharacter}
+          />
+          <CharacterPreviewScoringTeammate
+            index={1}
+            result={simScoringResult}
+            setCharacterModalOpen={setCharacterModalOpen}
+            setSelectedTeammateIndex={setSelectedTeammateIndex}
+            setCharacterModalInitialCharacter={setCharacterModalInitialCharacter}
+          />
+          <CharacterPreviewScoringTeammate
+            index={2}
+            result={simScoringResult}
+            setCharacterModalOpen={setCharacterModalOpen}
+            setSelectedTeammateIndex={setSelectedTeammateIndex}
+            setCharacterModalInitialCharacter={setCharacterModalInitialCharacter}
+          />
+        </Flex>
+      </Card>
+      <ScoreFooter result={simScoringResult}/>
+    </Flex>
+  )
+}
+
+function CharacterPreviewScoringTeammate(props: {
+  index: number
+  result: SimulationScore
+  setCharacterModalOpen: (b: boolean) => void
+  setSelectedTeammateIndex: (i: number | undefined) => void
+  setCharacterModalInitialCharacter: (character: Character) => void
+}) {
+  const { t } = useTranslation(['charactersTab', 'modals', 'common'])
+  const { result, index, setCharacterModalOpen, setSelectedTeammateIndex, setCharacterModalInitialCharacter } = props
+
+  const teammate = result.simulationMetadata.teammates[index]
+  const iconSize = 64
+
+  return (
+    <Card.Grid
+      style={{
+        width: '33.3333%',
+        textAlign: 'center',
+        padding: 1,
+        boxShadow: 'none',
+      }}
+      hoverable={true}
+      onClick={() => {
+        setCharacterModalInitialCharacter({ form: teammate } as Character)
+        setCharacterModalOpen(true)
+
+        setSelectedTeammateIndex(index)
+      }}
+      className='custom-grid'
+    >
+      <Flex vertical align='center' gap={0}>
+        <img
+          src={Assets.getCharacterAvatarById(teammate.characterId)}
+          style={{
+            height: iconSize,
+            width: iconSize,
+            borderRadius: 40,
+            background: 'rgba(255, 255, 255, 0.1)',
+            boxShadow: '0 4px 30px rgba(0, 0, 0, 0.2)',
+            backdropFilter: 'blur(5px)',
+            outline: '1px solid rgba(255, 255, 255, 0.3)',
+          }}
+        />
+        <OverlayText text={t('common:EidolonNShort', { eidolon: teammate.characterEidolon })} top={-12}/>
+        <img src={Assets.getLightConeIconById(teammate.lightCone)} style={{ height: iconSize, marginTop: -3 }}/>
+        <OverlayText
+          text={t('common:SuperimpositionNShort', { superimposition: teammate.lightConeSuperimposition })}
+          top={-18}
+        />
+      </Flex>
+    </Card.Grid>
+  )
+}
+
+export function ShowcaseDpsScoreHeader(props: {
+  relics: SingleRelicByPart
+  result: SimulationScore
+}) {
+  const { t } = useTranslation(['charactersTab'])
+
+  const result = props.result
+  const verified = Object.values(props.relics).filter((x) => x?.verified).length == 6
+
+  const textStyle: CSSProperties = {
+    fontSize: 17,
+    fontWeight: '600',
+    textAlign: 'center',
+    color: 'rgb(225, 165, 100)',
+    height: 23,
+    whiteSpace: 'nowrap',
+  }
+
+  const textDisplay = (
+    <Flex align='center' vertical style={{ marginBottom: 4, paddingTop: 3, paddingBottom: 3 }}>
+      <StatText style={textStyle}>
+        {t('CharacterPreview.ScoreHeader.Title')/* Combat Sim */}
+      </StatText>
+      <StatText style={textStyle}>
+        {
+          t(
+            'CharacterPreview.ScoreHeader.Score',
+            {
+              score: Utils.truncate10ths(Math.max(0, result.percent * 100)).toFixed(1),
+              grade: getSimScoreGrade(result.percent, verified),
+            },
+          )
+          /* DPS Score {{score}}% {{grade}} */
+        }
+      </StatText>
+    </Flex>
+  )
+
+  return (
+    <Flex vertical>
+      {textDisplay}
+    </Flex>
+  )
+}
+
+export function ScoreFooter(props: {
+  characterId: string
+  teamSelection: string
+  selectedTeammateIndex: number
+  characterModalInitialCharacter: Character | undefined
+  setTeamSelection: (teamSelection: string) => void
+  setRedrawTeammates: (random: number) => void
+}) {
+  const { t } = useTranslation(['charactersTab', 'modals', 'common'])
+  const [isCharacterModalOpen, setCharacterModalOpen] = useState(false)
+
+  const {
+    characterId,
+    teamSelection,
+    selectedTeammateIndex,
+    characterModalInitialCharacter,
+    setTeamSelection,
+    setRedrawTeammates,
+  } = props
+
+  // Teammate character modal OK
+  function onCharacterModalOk(form: Form) {
+    if (!form.characterId) {
+      return Message.error(t('CharacterPreview.Messages.NoSelectedCharacter')/* No selected character */)
+    }
+    if (!form.lightCone) {
+      return Message.error(t('CharacterPreview.Messages.NoSelectedCharacter')/* No selected character */)
+    }
+
+    const scoringMetadata = Utils.clone(DB.getScoringMetadata(characterId))
+    const simulation = scoringMetadata.simulation
+
+    simulation.teammates[selectedTeammateIndex] = form
+
+    DB.updateSimulationScoreOverrides(characterId, simulation)
+    setRedrawTeammates(Math.random())
+
+    setTeamSelection(CUSTOM_TEAM)
+  }
+
+  const tabsDisplay = (
+    <Segmented
+      style={{ marginLeft: 10, marginRight: 10, marginTop: 2, marginBottom: 0, alignItems: 'center' }}
+      onChange={(selection) => {
+        if (selection == SETTINGS_TEAM) {
+          window.modalApi.info({
+            icon: null,
+            width: 400,
+            okText: t('common:Ok'),
+            maskClosable: true,
+            content: (
+              <div style={{ width: '100%' }}>
+                <Flex vertical gap={10}>
+                  <HeaderText>{t('modals:ScoreFooter.ModalTitle')/* Combat sim scoring settings */}</HeaderText>
+                  <Button
+                    icon={<SyncOutlined/>}
+                    onClick={() => {
+                      const characterMetadata = Utils.clone(DB.getMetadata().characters[characterId])
+                      const simulation = characterMetadata.scoringMetadata.simulation
+
+                      DB.updateSimulationScoreOverrides(characterId, simulation)
+
+                      setTeamSelection(DEFAULT_TEAM)
+                      setRedrawTeammates(Math.random())
+
+                      Message.success(t('modals:ScoreFooter.ResetSuccessMsg')/* Reset to default teams */)
+                    }}
+                  >
+                    {t('modals:ScoreFooter.ResetButtonText')/* Reset custom team to default */}
+                  </Button>
+                  <Button
+                    icon={<SwapOutlined/>}
+                    onClick={() => {
+                      const characterMetadata = Utils.clone(DB.getScoringMetadata(characterId))
+                      const simulation = characterMetadata.simulation
+
+                      for (const teammate of simulation.teammates) {
+                        const form = DB.getCharacterById(teammate.characterId)?.form
+                        if (form == null) continue
+
+                        teammate.characterEidolon = form.characterEidolon
+                        if (form.lightCone) {
+                          teammate.lightCone = form.lightCone
+                          teammate.lightConeSuperimposition = form.lightConeSuperimposition || 1
+                        }
+                      }
+
+                      DB.updateSimulationScoreOverrides(characterId, simulation)
+                      setTeamSelection(CUSTOM_TEAM)
+                      setRedrawTeammates(Math.random())
+
+                      Message.success(t('modals:ScoreFooter.SyncSuccessMsg')/* Synced teammates */)
+                    }}
+                  >
+                    {t('modals:ScoreFooter.SyncButtonText')/* Sync imported eidolons / light cones */}
+                  </Button>
+                </Flex>
+              </div>
+            ),
+          })
+        } else {
+          setTeamSelection(selection)
+        }
+      }}
+      value={teamSelection}
+      block
+      options={[
+        {
+          value: DEFAULT_TEAM,
+          label: t('modals:ScoreFooter.TeamOptions.Default')/* Default */,
+        },
+        {
+          label: (
+            <SettingOutlined/>
+          ),
+          value: SETTINGS_TEAM,
+          className: 'short-segmented',
+        },
+        {
+          value: CUSTOM_TEAM,
+          label: t('modals:ScoreFooter.TeamOptions.Custom')/* Custom */,
+        },
+      ]}
+    />
+  )
+
+  return (
+    <Flex vertical style={{}} gap={2}>
+      <CharacterModal
+        onOk={onCharacterModalOk}
+        open={isCharacterModalOpen}
+        setOpen={setCharacterModalOpen}
+        initialCharacter={characterModalInitialCharacter}
+      />
+
+      {tabsDisplay}
+    </Flex>
+  )
+}
